Move configurable property attribute to own file
